feat: integrate chat page into main app; Co-authored by: Lukas Varga <lukas.varga128@gmail.com>
