--- src/frontend/TO INTEGRATE - Chat UI/app/_layout.tsx
@@ -1,9 +1,0 @@
-import { Stack } from "expo-router";
-
-export default function RootLayout() {
-  return (
-    <Stack>
-      <Stack.Screen name="index" text="Chat"/>
-    </Stack>
-  );
-}
